Move tokenization to a separate file.
